setup.py: - add check for cmake, - bugfix find_packages('src') - fix URL versioneer.py: bugfix: check file presence before unlink
ace-evaluator/CMakeLists.txt: remove extra message

bin: remove pace_update_ace, pace_update_yaml_potential

+++ versioneer.py
@@ -1354,7 +1354,8 @@ def versions_from_file(filename):
 
 def write_to_version_file(filename, versions):
     """Write the given version number to the given _version.py file."""
-    os.unlink(filename)
+    if os.path.isfile(filename):
+        os.unlink(filename)
     contents = json.dumps(versions, sort_keys=True,
                           indent=1, separators=(",", ": "))
     with open(filename, "w") as f:
